refactor: dublicate getProjectFolderName method
