+++ lib/commands/init/ui.js
@@ -6,6 +6,7 @@ const CONSTANTS = require('@src/utils/constants');
 const stringUtils = require('@src/utils/string-utils');
 const JsonView = require('@src/view/json-view');
 const Messenger = require('@src/view/messenger');
+const prompts = require('@src/view/prompt-view');
 
 module.exports = {
     showInitInstruction,
@@ -15,7 +16,7 @@ module.exports = {
     getCodeSrcForRegion,
     getSkillInfra,
     showPreviewAndConfirm,
-    getProjectFolderName
+    getProjectFolderName: prompts.getProjectFolderName
 };
 
 function showInitInstruction(profile) {
@@ -152,30 +153,6 @@ ${JsonView.toString(askStates)}
         name: 'confirmPreview'
     }]).then((answer) => {
         callback(null, answer.confirmPreview);
-    }).catch((error) => {
-        callback(error);
-    });
-}
-
-/**
- * To get user's input project folder name
- * @param {string} defaultName a default project name
- * @param {callback} callback { error, response }
- */
-function getProjectFolderName(defaultName, callback) {
-    inquirer.prompt([{
-        message: 'Please type in your folder name for the skill project (alphanumeric): ',
-        type: 'input',
-        default: defaultName,
-        name: 'projectFolderName',
-        validate: (input) => {
-            if (!input || stringUtils.filterNonAlphanumeric(input) === '') {
-                return 'Project folder name should consist of alphanumeric character(s) plus "-" only.';
-            }
-            return true;
-        }
-    }]).then((answer) => {
-        callback(null, stringUtils.filterNonAlphanumeric(answer.projectFolderName));
     }).catch((error) => {
         callback(error);
     });
