+++ lib/commands/new/ui.js
@@ -5,6 +5,7 @@ const inquirer = require('inquirer');
 
 const CONSTANTS = require('@src/utils/constants');
 const stringUtils = require('@src/utils/string-utils');
+const prompts = require('@src/view/prompt-view');
 
 const SKIP_DEPLOY_DELEGATE_SELECTION = 'self-hosted and manage your own hosting';
 
@@ -12,7 +13,7 @@ module.exports = {
     getSkillName,
     getSkillLocale,
     getSkillDefaultRegion,
-    getProjectFolderName,
+    getProjectFolderName: prompts.getProjectFolderName,
     selectSkillCodeLanguage,
     getTargetTemplateName,
     confirmUsingUnofficialTemplate,
@@ -63,25 +64,6 @@ function getSkillDefaultRegion(callback) {
         name: 'region'
     }]).then((answer) => {
         callback(null, CONSTANTS.HOSTED_SKILL.REGIONS[answer.region]);
-    }).catch((error) => {
-        callback(error);
-    });
-}
-
-function getProjectFolderName(defaultName, callback) {
-    inquirer.prompt([{
-        message: 'Please type in your folder name for the skill project (alphanumeric): ',
-        type: 'input',
-        default: defaultName,
-        name: 'projectFolderName',
-        validate: (input) => {
-            if (!input || stringUtils.filterNonAlphanumeric(input) === '') {
-                return 'Project folder name should be consisted of alphanumeric character(s) plus "-" only.';
-            }
-            return true;
-        }
-    }]).then((answer) => {
-        callback(null, stringUtils.filterNonAlphanumeric(answer.projectFolderName));
     }).catch((error) => {
         callback(error);
     });
