+++ test/unit/commands/new/ui-test.js
@@ -213,91 +213,6 @@ describe('Commands new - UI test', () => {
         });
     });
 
-    describe('# validate ui.getProjectFolderName', () => {
-        beforeEach(() => {
-            sinon.stub(inquirer, 'prompt');
-        });
-
-        afterEach(() => {
-            sinon.restore();
-        });
-
-        it('| project folder name is retrieved correctly', (done) => {
-            // setup
-            inquirer.prompt.resolves({ projectFolderName: TEST_FOLDER_NAME });
-            // call
-            ui.getProjectFolderName(TEST_REPO_NAME, (err, response) => {
-                // verify
-                validateInquirerConfig(inquirer.prompt.args[0][0][0], {
-                    message: 'Please type in your folder name for the skill project (alphanumeric): ',
-                    type: 'input',
-                    default: TEST_REPO_NAME,
-                });
-                expect(err).equal(null);
-                expect(response).equal(TEST_FOLDER_NAME);
-                done();
-            });
-        });
-
-        it('| project folder name is filtered correctly', (done) => {
-            // setup
-            const TEST_FOLDER_NAME_WITH_NON_ALPHANUMERIC = `${TEST_FOLDER_NAME}?/.%^&*`;
-            inquirer.prompt.resolves({ projectFolderName: TEST_FOLDER_NAME_WITH_NON_ALPHANUMERIC });
-            // call
-            ui.getProjectFolderName(TEST_REPO_NAME, (err, response) => {
-                // verify
-                validateInquirerConfig(inquirer.prompt.args[0][0][0], {
-                    message: 'Please type in your folder name for the skill project (alphanumeric): ',
-                    type: 'input',
-                    default: TEST_REPO_NAME,
-                });
-                expect(err).equal(null);
-                expect(response).equal(TEST_FOLDER_NAME);
-                done();
-            });
-        });
-
-        it('| get project folder name throws error from inquirer', (done) => {
-            // setup
-            inquirer.prompt.rejects(new Error(TEST_ERROR));
-            // call
-            ui.getProjectFolderName(TEST_REPO_NAME, (err, response) => {
-                // verify
-                validateInquirerConfig(inquirer.prompt.args[0][0][0], {
-                    message: 'Please type in your folder name for the skill project (alphanumeric): ',
-                    type: 'input',
-                    default: TEST_REPO_NAME,
-                });
-                expect(err.message).equal(TEST_ERROR);
-                expect(response).equal(undefined);
-                done();
-            });
-        });
-
-        it('| check the validate logic from inquirer and returns true', (done) => {
-            // setup
-            inquirer.prompt.resolves({ skillName: TEST_SKILL_NAME });
-            // call
-            ui.getProjectFolderName(TEST_REPO_NAME, () => {
-                // verify
-                expect(inquirer.prompt.args[0][0][0].validate('  !?!**  '))
-                    .equal('Project folder name should be consisted of alphanumeric character(s) plus "-" only.');
-                done();
-            });
-        });
-
-        it('| check the validate logic from inquirer and returns error', (done) => {
-            // setup
-            inquirer.prompt.resolves({ skillName: TEST_SKILL_NAME });
-            // call
-            ui.getProjectFolderName(TEST_REPO_NAME, () => {
-                // verify
-                expect(inquirer.prompt.args[0][0][0].validate('input')).equal(true);
-                done();
-            });
-        });
-    });
-
     describe('# validate ui.selectSkillCodeLanguage', () => {
         beforeEach(() => {
             sinon.stub(inquirer, 'prompt');
